# openapi.json
{"openapi":"3.1.0","info":{"title":"IP Prefix Lookup API","description":"API for performing IP lookups against cloud provider prefixes and managing prefix updates.","version":"1.0.0"},"paths":{"/api/v1/lookup/{ip}":{"get":{"tags":["IP Lookup"],"summary":"Get Lookup","description":"🔍 Lookup a single IP using Trie from FastAPI state.\n\n- **Validates IP** format before lookup.\n- **Returns** matching subnets, provider details, and tags.\n\nExample Request:\n```\nGET /api/v1/lookup/192.168.1.5\n```\n\nExample Response:\n```json\n{\n    \"result\": [\n        {\n            \"subnet\": \"192.168.1.0/24\",\n            \"provider\": \"AWS\",\n            \"tags\": [\"Cloud\"]\n        }\n    ]\n}\n```","operationId":"get_lookup_api_v1_lookup__ip__get","parameters":[{"name":"ip","in":"path","required":true,"schema":{"type":"string","format":"ipvanyaddress","title":"Ip"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/lookup/batch":{"post":{"tags":["IP Lookup"],"summary":"Post Lookup","description":"Perform a batch lookup for multiple IPs using Trie from FastAPI state.\n\n- **Validates IPs** before lookup.\n- **Returns a dictionary of matching results**.\n\nExample Request:\n```json\n{\n    \"ips\": [\"192.168.1.1\", \"2001:db8::ff00:42\"]\n}\n```\n\nExample Response:\n```json\n{\n    \"result\": {\n        \"192.168.1.1\": [\n            { \"subnet\": \"192.168.1.0/24\", \"provider\": \"AWS\", \"tags\": [\"Cloud\"] }\n        ],\n        \"2001:db8::ff00:42\": [\n            { \"subnet\": \"2001:db8::/32\", \"provider\": \"IPv6Provider\", \"tags\": [\"IPv6\"] }\n        ]\n    }\n}\n```","operationId":"post_lookup_api_v1_lookup_batch_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchIPRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reload/":{"post":{"tags":["Prefix reload"],"summary":"Reload Prefixes","description":"Endpoint to manually reload the Trie from `prefixes.json`.\n\nThis function:\n- Reads the latest `prefixes.json` file.\n- Updates the Compressed Radix Trie with new prefix data.\n- Logs every reload request for tracking.\n\nReturns:\n    JSON response confirming successful reload or an error message.","operationId":"reload_prefixes_api_v1_reload__post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/":{"get":{"tags":["Health Check"],"summary":"API Health Check","description":"Health check endpoint to verify that the API is running.\n\n- **Returns**: A simple message confirming the service is operational.\n- **Example Response**:\n```json\n{\n    \"message\": \"IP Prefix Lookup API is running.\"\n}\n```","operationId":"health_check_api_v1__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BatchIPRequest":{"properties":{"ips":{"items":{"type":"string","format":"ipvanyaddress"},"type":"array","minItems":1,"title":"IP List","description":"List of valid IPv4/IPv6 addresses."}},"type":"object","required":["ips"],"title":"BatchIPRequest","description":"Schema for batch IP lookup request.\n\n- `ips`: List of valid IPv4/IPv6 addresses.\n- Ensures at least one IP is provided."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}